rename async interceptors to have regular names & move sync intercept to interface
