--- src/Routine/Interception/AsyncInterceptorBase.cs
@@ -1,12 +1,0 @@
-﻿using Routine.Core.Runtime;
-
-namespace Routine.Interception;
-
-public abstract class AsyncInterceptorBase<TContext> : IInterceptor<TContext>
-    where TContext : InterceptionContext
-{
-    protected abstract Task<object> InterceptAsync(TContext context, Func<Task<object>> invocation);
-
-    object IInterceptor<TContext>.Intercept(TContext context, Func<object> invocation) => InterceptAsync(context, () => Task.FromResult(invocation())).WaitAndGetResult();
-    async Task<object> IInterceptor<TContext>.InterceptAsync(TContext context, Func<Task<object>> invocation) => await InterceptAsync(context, invocation);
-}
