+++ src/Routine/Interception/IInterceptor.cs
@@ -1,8 +1,11 @@
-﻿namespace Routine.Interception;
+﻿using Routine.Core.Runtime;
+
+namespace Routine.Interception;
 
 public interface IInterceptor<in TContext>
     where TContext : InterceptionContext
 {
-    object Intercept(TContext context, Func<object> invocation);
     Task<object> InterceptAsync(TContext context, Func<Task<object>> invocation);
+
+    public object Intercept(TContext context, Func<object> invocation) => InterceptAsync(context, () => Task.FromResult(invocation())).WaitAndGetResult();
 }
