+++ src/Routine/Interception/LastChainLinkInterceptor.cs
@@ -5,6 +5,5 @@ public class LastChainLinkInterceptor<TContext> : IChainLinkInterceptor<TContext
 {
     public IChainLinkInterceptor<TContext> Next { get => null; set { } }
 
-    public object Intercept(TContext context, Func<object> invocation) => invocation();
     public async Task<object> InterceptAsync(TContext context, Func<Task<object>> invocation) => await invocation();
 }
